+++ pkg/integrations/v4/emitter/emitter_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/newrelic/infrastructure-agent/internal/agent/mocks"
 
 	"github.com/newrelic/infrastructure-agent/internal/agent"
-	"github.com/newrelic/infrastructure-agent/internal/agent/cmdchannel/handler"
+	"github.com/newrelic/infrastructure-agent/internal/agent/cmdchannel/ffhandler"
 	"github.com/newrelic/infrastructure-agent/internal/feature_flags"
 	"github.com/newrelic/infrastructure-agent/internal/integrations/v4/integration"
 	"github.com/newrelic/infrastructure-agent/pkg/config"
@@ -431,7 +431,7 @@ func TestLegacy_Emit(t *testing.T) {
 
 			em := &VersionAwareEmitter{
 				aCtx:        ma,
-				ffRetriever: feature_flags.NewManager(map[string]bool{handler.FlagProtocolV4: true, handler.FlagDMRegisterEnable: true}),
+				ffRetriever: feature_flags.NewManager(map[string]bool{ffhandler.FlagProtocolV4: true, ffhandler.FlagDMRegisterEnable: true}),
 				dmEmitter:   mockDME,
 			}
 
@@ -469,7 +469,7 @@ func TestProtocolV4_Emit(t *testing.T) {
 
 	em := &VersionAwareEmitter{
 		aCtx:        ma,
-		ffRetriever: feature_flags.NewManager(map[string]bool{handler.FlagProtocolV4: true, handler.FlagDMRegisterEnable: true}),
+		ffRetriever: feature_flags.NewManager(map[string]bool{ffhandler.FlagProtocolV4: true, ffhandler.FlagDMRegisterEnable: true}),
 		dmEmitter:   mockDME,
 	}
 
@@ -527,7 +527,7 @@ func TestProtocolV4_Emit_WithFFDisabled(t *testing.T) {
 
 	em := &VersionAwareEmitter{
 		aCtx:        ma,
-		ffRetriever: feature_flags.NewManager(map[string]bool{handler.FlagProtocolV4: false, handler.FlagDMRegisterEnable: true}),
+		ffRetriever: feature_flags.NewManager(map[string]bool{ffhandler.FlagProtocolV4: false, ffhandler.FlagDMRegisterEnable: true}),
 		dmEmitter:   mockDME,
 	}
 
@@ -555,7 +555,7 @@ func TestProtocolV4_Emit_WithoutRegisteringEntities(t *testing.T) {
 
 	em := &VersionAwareEmitter{
 		aCtx:        mockAgent(),
-		ffRetriever: feature_flags.NewManager(map[string]bool{handler.FlagProtocolV4: true, handler.FlagDMRegisterEnable: false}),
+		ffRetriever: feature_flags.NewManager(map[string]bool{ffhandler.FlagProtocolV4: true, ffhandler.FlagDMRegisterEnable: false}),
 		dmEmitter:   dmEmitter,
 	}
